DatingApp/requirement.txt
rest_framework
django-rest-knox
